ADD SCHOLARSHIP AND REWORK CODE
Add the virtual opportunities scholarship. This was fine, but I realised
that the code really didn't achieve everything it was meant to properly,
cleanly. So, restructure everything. Make the tables environment work
properly using tabularx. Edit some of the commands and the way they play
together to make the tables work as expected.