ADD PUBLICATION SUBMISSIONS AND EXTRACURRICULAR
Add a command for publications, which handles their formatting in a very
basic way. In future, it would be good to:
- maintain a bib file of all my publications
- define \pub{} with bibtex / biblatex as a reference
- define the publications section just as \pub references to EVERYTHING in
  the bib file.
(these notes are also in the cv.tex itself)

Using this command, add publications / submissions section with
submissions to Aporia and Prokopton.

Add extracurricular section with FSC and sailing.