amend wording on teaching evals intro
In the recent feedback, one student said I was handsome. It made me
feel icky and I decided to not include that record. So it's no longer
true that I include *all* comments, and I don't want to represent
myself as such. On the other hand, drawing attention to the fact that
I have chosen some comments and not others might make it seem as if I
editorialise, which I don't. I've decided this is a good middle
ground (i.e. wording it so that the issue does not arise!).